% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addline_format.R
\name{addline_format}
\alias{addline_format}
\title{Plot multi-line axis labels}
\usage{
addline_format(x, ...)
}
\description{
\code{addline_format} Multi-line axis labels
}
\examples{
# 
ggplot(counts, aes(x=norm, y=value, fill=type)) +
geom_col() +
scale_x_discrete(labels=addline_format(c("FirstLine SecondLine", "Two-line Label")))

}
