% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.to.1.R
\name{rescale.to.1}
\alias{rescale.to.1}
\title{Rescales a variable to 1}
\usage{
rescale.to.1(var, num)
}
\arguments{
\item{var}{a vector to be rescaled}

\item{num}{max number on current scale (for a 7-point scale, num=7). Minimum must be 1}
}
\description{
Rescales a variable to 1
}
\examples{
rescale.to.1()
}
