% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustse.f.R
\name{robustse.f}
\alias{robustse.f}
\title{Calculates Huber-White heteroskedasticity-robust standard errors}
\usage{
robustse.f(model, cluster, df_correction)
}
\arguments{
\item{model}{an lm or glm object}

\item{cluster}{the variable to be clustered, prepended by a ~ (e.g., ~state).}

\item{df_correction}{Setting df_correction = F will not count levels of your cluster variable against degrees of freedom.}
}
\description{
Calculates Huber-White heteroskedasticity-robust standard errors
}
\examples{
robustse.f()
}
\keyword{errors}
\keyword{standard}
