% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardizes x from 0 to 1}
\usage{
standardize(x, na.rm = T)
}
\arguments{
\item{x}{a vector to be standardized}
}
\description{
Standardizes x from 0 to 1
}
\examples{
standardize()
}
\keyword{cleaning}
\keyword{data}
