% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_james.R
\name{theme_james}
\alias{theme_james}
\title{James's ggplot2 theme}
\usage{
theme_james(...)
}
\arguments{
\item{...}{Passed to [ggplot2::theme()]}

\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
An object as returned by [ggplot2::theme()]
}
\description{
Code based on theme_karl: https://github.com/kbroman/broman/blob/master/R/theme_karl.R
and theme_unicorn: https://www.r-bloggers.com/how-to-create-a-ggplot-theme-unicorn-edition/
}
\examples{
library(ggplot2)
mtcars$cyl <- factor(mtcars$cyl)
ggplot(mtcars, aes(y=mpg, x=disp, color=cyl)) +
    geom_point() + theme_james()

}
\seealso{
[ggplot2::theme()]
}
