% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_code.R
\name{reverse_code}
\alias{reverse_code}
\title{Reverse Codes a variable}
\usage{
reverse_code(var)
}
\arguments{
\item{var}{a numeric vector to be reverse coded}
}
\value{
a reverse coded vector
standardize()
}
\description{
Reverse Codes a variable
}
\examples{
my_variable <- c(1,2,2,3,3,1)
new_variable <- reverse_code (var=my_variable)
new_variable
}
\keyword{cleaning}
\keyword{data}
